Cleaned up example source code files
